Fix console errors
Here's some changes that don't affect the functionality of the application, but clean up some errors in the developer console (by adopting React recommended practices).

* JSX requires `className` instead of class
* JSX components returned by map should have a `key` property. Internally, React uses this in order to keep track of which items are the same when it calls the functions again to re-render the view when something changes.

This commit also contains some changes to the spaces, just because I have my editor set to format JavaScript when a file is saved and I didn't disable that feature for making these changes.